Fix longest break duration
If you do an activity say on Monday and on Wednesday, then there is a break
of one day where you didn't do that activity (longest break = 1 day). Hence
if you do an activity on two consecutive days, then there is no break at all
(longest break = 0 days). Currently Tickmate shows 2 and 1 days, respectively,
in the above cases. This is going to be fixed by this commit.