Additional option for editable days: today, yesterday/last weekend
This is a modification of the today + yesterday (active day and day before) option
for the case that you'll update your diary on weekdays only. On Mondays it additionally
allows the preceding Saturday so that you can update activities that happened on last
weekend.
Localizations for PL, CZ, and IT yet to be supplied.