Time info code cleanup
Time info is saved to the database for ticks on the current day but not used in the program
anywhere. On the contrary, all calendar operations rely on the fact that the time info is
cleared. This can be simplified by not loading the time info from the database in the first place.
This allows to remove the instructions clearing the hour, minutes, seconds, and milliseconds
scattered all over the code.