Fix crash on preference dialog
Fixes #162

Problem is that the EditText of an EditTextPreference is not part of the
layout when setting a custom layout. But the EditText is used like it is
part of the layout and will return null every for various calls that
will lead to a crash. By adding the EditText to the custom layout, it
works. But it's still just a hack.